refactor: :recycle: Component refactored
This commits moves the component AverageBarChat to the folder components of the AverageBarChartCard
