+++ fe/src/view/pages/Dashboard/components/AverageBarChartCard/components/AverageBarChart/index.tsx
